Move isComp check into Robot class as a static so all constants files can reference it
