+++ Comp2024/src/main/java/frc/robot/lib/util/CTREConfigs6.java
@@ -9,6 +9,7 @@ import com.ctre.phoenix6.signals.NeutralModeValue;
 import com.ctre.phoenix6.signals.SensorDirectionValue;
 
 import frc.robot.Constants.Ports;
+import frc.robot.Robot;
 import frc.robot.lib.math.Conversions;
 
 public final class CTREConfigs6
@@ -92,7 +93,10 @@ public final class CTREConfigs6
 
     // config.MagnetSensor.AbsoluteSensorRange = AbsoluteSensorRangeValue.Signed_PlusMinusHalf;
     // config.MagnetSensor.SensorDirection = SWConsts.steerCanCoderInvert;
-    // // config.MagnetSensor.MagnetOffset 
+    // if (Robot.isReal( ))
+    //   config.MagnetSensor.MagnetOffset = (Robot.isComp( )) ? 0.0 : 0.0;
+    // else
+    //   config.MagnetSensor.MagnetOffset = -0.25; // Simulated CANcoder default
 
     return config;
   }
@@ -170,7 +174,10 @@ public final class CTREConfigs6
 
     config.MagnetSensor.AbsoluteSensorRange = AbsoluteSensorRangeValue.Signed_PlusMinusHalf;
     config.MagnetSensor.SensorDirection = SensorDirectionValue.Clockwise_Positive;
-    config.MagnetSensor.MagnetOffset = -0.891113;
+    if (Robot.isReal( ))
+      config.MagnetSensor.MagnetOffset = (Robot.isComp( )) ? 0.0 : -0.891113; // TODO: get comp value
+    else
+      config.MagnetSensor.MagnetOffset = -0.25; // Simulated CANcoder default
 
     return config;
   }
